requirements.txt
Python 3.10
py-solc-x
web3
python-dotenv